Split up `chatEditingService.ts`
